move SPNetworkClient to its own file
